Throw error when `seed` or `schema` are undefined
Throw error when the `seed` or `schema` parameter for the
account object are `undefined`. This will make it more
obvious when the parameters are missing the using the
package stalls.

When using the package it is recommended to use a `.env` file
to load the parameters from. This error will help identify
when the `.env` is probably missing.